requirements.txt
python-dotenv==0.20.0
streamlit==1.35.0
openai==1.30.5
extra-streamlit-components==0.1.71
streamlit-authenticator==0.3.2

